Refactor ExpertResultCard and GrantInvitedExpertCard components to utilize the new ExpertSourceLinkIcon component for improved source link rendering.
